Moved almost all style generation to the map-style file
